Step: new parent: move
Remove deprecated test class
