+++ parallel-consumer-core/src/main/java/io/confluent/parallelconsumer/JStreamParallelStreamProcessor.java
@@ -3,6 +3,7 @@ package io.confluent.parallelconsumer;
 /*-
  * Copyright (C) 2020-2021 Confluent, Inc.
  */
+import io.confluent.parallelconsumer.internal.AbstractParallelEoSStreamProcessor;
 import io.confluent.parallelconsumer.internal.DrainingCloseable;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.apache.kafka.clients.producer.ProducerRecord;
@@ -18,7 +19,7 @@ public interface JStreamParallelStreamProcessor<K, V> extends DrainingCloseable 
     }
 
     /**
-     * Like {@link ParentParallelEoSStreamProcessor#pollAndProduceMany} but instead of callbacks, streams the results
+     * Like {@link AbstractParallelEoSStreamProcessor#pollAndProduceMany} but instead of callbacks, streams the results
      * instead, after the produce result is ack'd by Kafka.
      *
      * @return a stream of results of applying the function to the polled records
