+++ parallel-consumer-core/src/main/java/io/confluent/parallelconsumer/ParallelConsumer.java
@@ -4,6 +4,7 @@ package io.confluent.parallelconsumer;
  * Copyright (C) 2020-2021 Confluent, Inc.
  */
 
+import io.confluent.parallelconsumer.internal.AbstractParallelEoSStreamProcessor;
 import io.confluent.parallelconsumer.internal.DrainingCloseable;
 import lombok.Data;
 import org.apache.kafka.clients.consumer.ConsumerRebalanceListener;
@@ -13,15 +14,16 @@ import java.util.function.Consumer;
 import java.util.regex.Pattern;
 
 // tag::javadoc[]
+
 /**
  * Asynchronous / concurrent message consumer for Kafka.
  * <p>
  * Currently there is no direct implementation, only the {@link ParallelStreamProcessor} version (see {@link
- * ParentParallelEoSStreamProcessor}), but there may be in the future.
+ * AbstractParallelEoSStreamProcessor}), but there may be in the future.
  *
  * @param <K> key consume / produce key type
  * @param <V> value consume / produce value type
- * @see ParentParallelEoSStreamProcessor
+ * @see AbstractParallelEoSStreamProcessor
  * @see #poll(Consumer)
  */
 // end::javadoc[]
