+++ parallel-consumer-core/src/test/java/io/confluent/csid/utils/LongPollingMockConsumer.java
@@ -4,7 +4,7 @@ package io.confluent.csid.utils;
  * Copyright (C) 2020-2021 Confluent, Inc.
  */
 
-import io.confluent.parallelconsumer.ParentParallelEoSStreamProcessor;
+import io.confluent.parallelconsumer.internal.AbstractParallelEoSStreamProcessor;
 import lombok.Getter;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
@@ -107,7 +107,7 @@ public class LongPollingMockConsumer<K, V> extends MockConsumer<K, V> {
      * assignments, use reflection to access the registered rebalance listener, call the listener, and only then close
      * the consumer.
      *
-     * @see ParentParallelEoSStreamProcessor#onPartitionsRevoked
+     * @see AbstractParallelEoSStreamProcessor#onPartitionsRevoked
      */
     private void revokeAssignment() throws NoSuchFieldException, IllegalAccessException {
         ConsumerRebalanceListener consumerRebalanceListener = getRebalanceListener();
