+++ parallel-consumer-core/src/test/java/io/confluent/parallelconsumer/ParallelEoSStreamProcessorTestBase.java
@@ -4,6 +4,7 @@ package io.confluent.parallelconsumer;
  * Copyright (C) 2020-2021 Confluent, Inc.
  */
 
+import io.confluent.parallelconsumer.internal.AbstractParallelEoSStreamProcessor;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
@@ -12,13 +13,13 @@ public class ParallelEoSStreamProcessorTestBase extends AbstractParallelEoSStrea
     protected ParallelEoSStreamProcessor<String, String> parallelConsumer;
 
     @Override
-    protected ParentParallelEoSStreamProcessor<String, String> initAsyncConsumer(ParallelConsumerOptions parallelConsumerOptions) {
-        parallelConsumer = initPollingAsyncConsumer(parallelConsumerOptions);
-        return parallelConsumer;
+    protected AbstractParallelEoSStreamProcessor<String, String> initAsyncConsumer(ParallelConsumerOptions parallelConsumerOptions) {
+        return initPollingAsyncConsumer(parallelConsumerOptions);
     }
 
     protected ParallelEoSStreamProcessor<String, String> initPollingAsyncConsumer(ParallelConsumerOptions parallelConsumerOptions) {
         parallelConsumer = new ParallelEoSStreamProcessor<>(parallelConsumerOptions);
+        super.parentParallelConsumer = parallelConsumer;
         return parallelConsumer;
     }
